refactor: reorganize wine-n-beer feature and update associated imports
- Relocated `drunk` logic and `alcohol` component to `features/wine-n-beer`.
- Renamed `load_alcohol_component` to `loadOnDrinkComponent` and `load_drunk` to `drunkDataManager`.
- Updated import paths across modules to reflect the new structure.
