+++ behaviour_pack/scripts-dev/components/index.ts
@@ -4,7 +4,7 @@ import load_altar_component from "./sacrifice";
 import load_whoop_component from "./whoop";
 import load_reactor_activate_component from "./reactor_activate";
 import load_heal_dragon_component from "../features/dragon-fight/draconic-heart";
-import load_alcohol_component from "./alcohol";
+import load_alcohol_component from "../features/wine-n-beer/components";
 
 export default function load_custom_components(guild_id: string) {
     // Block Components
