+++ behaviour_pack/scripts-dev/features/wine-n-beer/index.ts
@@ -1,0 +1,7 @@
+import loadOnDrinkComponent from "./components";
+import drunkDataManager from "./loops";
+
+export default function loadWineAndBeerFeature() {
+    loadOnDrinkComponent()
+    drunkDataManager()
+}
